packetizer: hevc: fix nal size storage

+++ modules/packetizer/hevc_nal.h
@@ -325,11 +325,11 @@ struct hevc_dcr_params
                   *p_pps[HEVC_DCR_VPS_COUNT],
                   *p_seipref[HEVC_DCR_SEI_COUNT],
                   *p_seisuff[HEVC_DCR_SEI_COUNT];
-    uint8_t rgi_vps[HEVC_DCR_VPS_COUNT],
-            rgi_sps[HEVC_DCR_SPS_COUNT],
-            rgi_pps[HEVC_DCR_PPS_COUNT],
-            rgi_seipref[HEVC_DCR_SEI_COUNT],
-            rgi_seisuff[HEVC_DCR_SEI_COUNT];
+    uint16_t rgi_vps[HEVC_DCR_VPS_COUNT],
+             rgi_sps[HEVC_DCR_SPS_COUNT],
+             rgi_pps[HEVC_DCR_PPS_COUNT],
+             rgi_seipref[HEVC_DCR_SEI_COUNT],
+             rgi_seisuff[HEVC_DCR_SEI_COUNT];
     uint8_t i_vps_count, i_sps_count, i_pps_count;
     uint8_t i_seipref_count, i_seisuff_count;
     struct hevc_dcr_values *p_values;
